framework_crewai/blog_crew/requirements-minimal.txt
# Versão mínima dos requirements para testes rápidos
crewai>=0.28.0
openai>=1.0.0
google-generativeai
python-dotenv
requests
Pillow
redis
pydantic>=2.0.0
sanity-client>=3.0.0